requirements.txt
streamlit==1.32.0
neo4j==5.15.0
langchain==0.1.12
langchain-groq==0.0.5
sentence-transformers==2.5.1
networkx==3.2.1
plotly==5.19.0
pyvis==0.4.1
python-dotenv==1.0.1
pypdf==4.0.2
